refactor useful functions to utils
