+++ bulkformer/__init__.py
@@ -8,10 +8,16 @@ from bulkformer.models.rope import PositionalExprEmbedding
 from bulkformer.models.encoder import GBFormer
 from bulkformer.models.model import BulkFormer
 
+# Import utility functions for data processing and feature extraction
+from bulkformer.utils import normalize_data, align_genes, extract_features
+
 __all__ = [
     "BulkFormer",
     "GBFormer",
     "PositionalExprEmbedding",
     "model_params",
+    "normalize_data",
+    "align_genes",
+    "extract_features",
 ]
