+++ bulkformer/utils/__init__.py
@@ -1,0 +1,16 @@
+"""
+Utility functions for BulkFormer.
+"""
+
+from bulkformer.utils.features import (
+    normalize_data,
+    align_genes,
+    extract_features,
+)
+
+__all__ = [
+    "normalize_data",
+    "align_genes", 
+    "extract_features",
+]
+
